requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
certifi==2025.8.3
charset-normalizer==3.4.3
disnake==2.10.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
idna==3.10
multidict==6.6.4
propcache==0.3.2
python-dotenv==1.1.1
requests==2.32.5
selectolax==0.3.33
sniffio==1.3.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
